增加用户权限管理

+++ src/main/java/com/swpu/uchain/openexperiment/service/UserRoleService.java
@@ -25,7 +25,7 @@ public interface UserRoleService {
      * @param userId
      * @param roleId
      */
-    Result deleteByUserIdRoleId(Long userId, Long roleId);
+    Result deleteByUserIdRoleId(Long userId, Integer roleId);
 
     /**
      * 为用户配置角色
